:necktie: LINT: Improve type hints and rename regions module
