+++ assignments/life_expectancy/regions.py
